run fmt

+++ contracts/callpaths/ColdPath.sol
@@ -2,20 +2,20 @@
 
 pragma solidity 0.8.19;
 
-import '../libraries/Directives.sol';
-import '../libraries/Encoding.sol';
-import '../libraries/TokenFlow.sol';
-import '../libraries/PriceGrid.sol';
-import '../libraries/ProtocolCmd.sol';
-import '../mixins/SettleLayer.sol';
-import '../mixins/PoolRegistry.sol';
-import '../mixins/MarketSequencer.sol';
-import '../mixins/StorageLayout.sol';
-import '../mixins/ProtocolAccount.sol';
-import '../mixins/DepositDesk.sol';
-import '../periphery/CrocLpErc20.sol';
-import '../interfaces/ICrocMinion.sol';
-import '../CrocEvents.sol';
+import "../libraries/Directives.sol";
+import "../libraries/Encoding.sol";
+import "../libraries/TokenFlow.sol";
+import "../libraries/PriceGrid.sol";
+import "../libraries/ProtocolCmd.sol";
+import "../mixins/SettleLayer.sol";
+import "../mixins/PoolRegistry.sol";
+import "../mixins/MarketSequencer.sol";
+import "../mixins/StorageLayout.sol";
+import "../mixins/ProtocolAccount.sol";
+import "../mixins/DepositDesk.sol";
+import "../periphery/CrocLpErc20.sol";
+import "../interfaces/ICrocMinion.sol";
+import "../CrocEvents.sol";
 
 /* @title Cold path callpath sidecar.
  * @notice Defines a proxy sidecar contract that's used to move code outside the 
@@ -34,11 +34,10 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
     using Chaining for Chaining.PairFlow;
     using ProtocolCmd for bytes;
 
-
-    constructor(address initialWbera) DepositDesk(initialWbera) {
-    }
+    constructor(address initialWbera) DepositDesk(initialWbera) {}
     /* @notice Consolidated method for protocol control related commands. */
-    function protocolCmd (bytes calldata cmd) public {
+
+    function protocolCmd(bytes calldata cmd) public {
         uint8 code = uint8(cmd[31]);
 
         if (code == ProtocolCmd.DISABLE_TEMPLATE_CODE) {
@@ -64,10 +63,9 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
         emit CrocEvents.CrocColdProtocolCmd(cmd);
     }
 
-    
-    function userCmd (bytes calldata cmd) public payable {
+    function userCmd(bytes calldata cmd) public payable {
         uint8 cmdCode = uint8(cmd[31]);
-        
+
         if (cmdCode == UserCmd.INIT_POOL_CODE) {
             initPool(cmd);
         } else if (cmdCode == UserCmd.APPROVE_ROUTER_CODE) {
@@ -94,16 +92,16 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
 
         emit CrocEvents.CrocColdCmd(cmd);
     }
-    
+
     /* @notice Initializes the pool type for the pair.
      * @param base The base token in the pair.
      * @param quote The quote token in the pair.
      * @param poolIdx The index of the pool type to initialize.
      * @param price The price to initialize the pool. Represented as square root price in
      *              Q64.64 notation. */
-    function initPool (bytes calldata cmd) private {
+    function initPool(bytes calldata cmd) private {
         (, address base, address quote, uint256 poolIdx, uint128 price) =
-            abi.decode(cmd, (uint8, address,address,uint256,uint128));
+            abi.decode(cmd, (uint8, address, address, uint256, uint128));
         bool nativeBera = false;
         if (base == address(0)) {
             base = wbera;
@@ -114,7 +112,7 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
         }
         (PoolSpecs.PoolCursor memory pool, uint128 initLiq) = registerPool(base, quote, poolIdx);
         verifyPermitInit(pool, base, quote, poolIdx);
-        
+
         (int128 baseFlow, int128 quoteFlow) = initCurve(pool, price, initLiq);
 
         if (nativeBera) {
@@ -141,12 +139,12 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
     /* @notice Disables an existing pool template. Any previously instantiated pools on
      *         this template will continue exist, but calling this will prevent any new
      *         pools from being created on this template. */
-    function disableTemplate (bytes calldata input) private {
+    function disableTemplate(bytes calldata input) private {
         (, uint256 poolIdx) = abi.decode(input, (uint8, uint256));
         emit CrocEvents.DisablePoolTemplate(poolIdx);
         disablePoolTemplate(poolIdx);
     }
-    
+
     /* @notice Sets template parameters for a pool type index.
      * @param poolIdx The index of the pool type.
      * @param feeRate The pool's swap fee rate in multiples of 0.0001%
@@ -157,44 +155,52 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
      * @param priceFloor The lower limit of price at which mint/burn of concentrated liq is allowed for stable swap pool (defined as square root of actual price).
      * @param priceCeiling The upper limit of price at which mint/burn of concentrated liq is allowed for stable swap pool (defined as square root of actual price).
      * @param isStableSwapPool The stable swap pool flag. */
-    function setTemplate (bytes calldata input) private {
-        (, uint256 poolIdx, uint16 feeRate, uint16 tickSize, uint8 jitThresh,
-         uint8 knockout, uint8 oracleFlags, uint128 priceFloor, uint128 priceCeiling, bool isStableSwapPool) =
-            abi.decode(input, (uint8, uint256, uint16, uint16, uint8, uint8, uint8, uint128, uint128, bool));
-        
-        emit CrocEvents.SetPoolTemplate(poolIdx, feeRate, tickSize, jitThresh, knockout,
-                                        oracleFlags, priceFloor, priceCeiling);
-        setPoolTemplate(poolIdx, feeRate, tickSize, jitThresh, knockout, oracleFlags, priceFloor, priceCeiling, isStableSwapPool);
+    function setTemplate(bytes calldata input) private {
+        (
+            ,
+            uint256 poolIdx,
+            uint16 feeRate,
+            uint16 tickSize,
+            uint8 jitThresh,
+            uint8 knockout,
+            uint8 oracleFlags,
+            uint128 priceFloor,
+            uint128 priceCeiling,
+            bool isStableSwapPool
+        ) = abi.decode(input, (uint8, uint256, uint16, uint16, uint8, uint8, uint8, uint128, uint128, bool));
+
+        emit CrocEvents.SetPoolTemplate(
+            poolIdx, feeRate, tickSize, jitThresh, knockout, oracleFlags, priceFloor, priceCeiling
+        );
+        setPoolTemplate(
+            poolIdx, feeRate, tickSize, jitThresh, knockout, oracleFlags, priceFloor, priceCeiling, isStableSwapPool
+        );
     }
 
-    function setTakeRate (bytes calldata input) private {
-        (, uint8 takeRate) = 
-            abi.decode(input, (uint8, uint8));
-        
+    function setTakeRate(bytes calldata input) private {
+        (, uint8 takeRate) = abi.decode(input, (uint8, uint8));
+
         emit CrocEvents.SetTakeRate(takeRate);
         setProtocolTakeRate(takeRate);
     }
 
-    function setRelayerTakeRate (bytes calldata input) private {
-        (, uint8 takeRate) = 
-            abi.decode(input, (uint8, uint8));
+    function setRelayerTakeRate(bytes calldata input) private {
+        (, uint8 takeRate) = abi.decode(input, (uint8, uint8));
 
         emit CrocEvents.SetRelayerTakeRate(takeRate);
         setRelayerTakeRate(takeRate);
     }
 
-    function setNewPoolLiq (bytes calldata input) private {
-        (, uint128 liq) = 
-            abi.decode(input, (uint8, uint128));
-        
+    function setNewPoolLiq(bytes calldata input) private {
+        (, uint128 liq) = abi.decode(input, (uint8, uint128));
+
         emit CrocEvents.SetNewPoolLiq(liq);
         setNewPoolLiq(liq);
     }
 
-    function resyncTakeRate (bytes calldata input) private {
-        (, address base, address quote, uint256 poolIdx) = 
-            abi.decode(input, (uint8, address, address, uint256));
-        
+    function resyncTakeRate(bytes calldata input) private {
+        (, address base, address quote, uint256 poolIdx) = abi.decode(input, (uint8, address, address, uint256));
+
         emit CrocEvents.ResyncTakeRate(base, quote, poolIdx, protocolTakeRate_);
         resyncProtocolTake(base, quote, poolIdx);
     }
@@ -210,10 +216,19 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
      * @param knockout The knockout bit flags for the pool. 
      * @param priceFloor The lower limit of price at which mint/burn of concentrated liq is allowed for stable swap pool (defined as square root of actual price).
      * @param priceCeiling The upper limit of price at which mint/burn of concentrated liq is allowed for stable swap pool (defined as square root of actual price). */
-    function revisePool (bytes calldata cmd) private {
-        (, address base, address quote, uint256 poolIdx,
-         uint16 feeRate, uint16 tickSize, uint8 jitThresh, uint8 knockout, uint128 priceFloor, uint128 priceCeiling) =
-            abi.decode(cmd, (uint8,address,address,uint256,uint16,uint16,uint8,uint8, uint128, uint128));
+    function revisePool(bytes calldata cmd) private {
+        (
+            ,
+            address base,
+            address quote,
+            uint256 poolIdx,
+            uint16 feeRate,
+            uint16 tickSize,
+            uint8 jitThresh,
+            uint8 knockout,
+            uint128 priceFloor,
+            uint128 priceCeiling
+        ) = abi.decode(cmd, (uint8, address, address, uint256, uint16, uint16, uint8, uint8, uint128, uint128));
         setPoolSpecs(base, quote, poolIdx, feeRate, tickSize, jitThresh, knockout, priceFloor, priceCeiling);
     }
 
@@ -222,7 +237,7 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
      * @param unitTickCollateral The collateral threshold for off-grid price improvement.
      * @param awayTickTol The maximum tick distance from current price that off-grid
      *                    quotes are allowed for. */
-    function pegPriceImprove (bytes calldata cmd) private {
+    function pegPriceImprove(bytes calldata cmd) private {
         (, address token, uint128 unitTickCollateral, uint16 awayTickTol) =
             abi.decode(cmd, (uint8, address, uint128, uint16));
         emit CrocEvents.PriceImproveThresh(token, unitTickCollateral, awayTickTol);
@@ -236,53 +251,46 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
      *              paid from the user into the pool, increasing their balance.
      * @param token The token to which the surplus collateral is applied. (If 0x0, then
      *              native Ethereum) */
-    function depositSurplus (bytes calldata cmd) private {
-        (, address recv, uint128 value, address token) =
-            abi.decode(cmd, (uint8, address, uint128, address));
+    function depositSurplus(bytes calldata cmd) private {
+        (, address recv, uint128 value, address token) = abi.decode(cmd, (uint8, address, uint128, address));
         depositSurplus(recv, value, token);
     }
 
-    function depositPermit (bytes calldata cmd) private {
-        (, address recv, uint128 value, address token, uint256 deadline,
-         uint8 v, bytes32 r, bytes32 s) =
-            abi.decode(cmd, (uint8, address, uint128, address, uint256,
-                             uint8, bytes32, bytes32));
+    function depositPermit(bytes calldata cmd) private {
+        (, address recv, uint128 value, address token, uint256 deadline, uint8 v, bytes32 r, bytes32 s) =
+            abi.decode(cmd, (uint8, address, uint128, address, uint256, uint8, bytes32, bytes32));
         depositSurplusPermit(recv, value, token, deadline, v, r, s);
     }
 
-    function disburseSurplus (bytes calldata cmd) private {
-        (, address recv, int128 value, address token) =
-            abi.decode(cmd, (uint8, address, int128, address));
+    function disburseSurplus(bytes calldata cmd) private {
+        (, address recv, int128 value, address token) = abi.decode(cmd, (uint8, address, int128, address));
         disburseSurplus(recv, value, token);
     }
 
-    function transferSurplus (bytes calldata cmd) private {
-        (, address recv, int128 size, address token) =
-            abi.decode(cmd, (uint8, address, int128, address));
+    function transferSurplus(bytes calldata cmd) private {
+        (, address recv, int128 size, address token) = abi.decode(cmd, (uint8, address, int128, address));
         transferSurplus(recv, size, token);
     }
 
-    function sidePocketSurplus (bytes calldata cmd) private {
+    function sidePocketSurplus(bytes calldata cmd) private {
         (, uint256 fromSalt, uint256 toSalt, int128 value, address token) =
             abi.decode(cmd, (uint8, uint256, uint256, int128, address));
         sidePocketSurplus(fromSalt, toSalt, value, token);
     }
 
-    function resetNonce (bytes calldata cmd) private {
-        (, bytes32 salt, uint32 nonce) = 
-            abi.decode(cmd, (uint8, bytes32, uint32));
+    function resetNonce(bytes calldata cmd) private {
+        (, bytes32 salt, uint32 nonce) = abi.decode(cmd, (uint8, bytes32, uint32));
         resetNonce(salt, nonce);
     }
-    
-    function resetNonceCond (bytes calldata cmd) private {
-        (, bytes32 salt, uint32 nonce, address oracle, bytes memory args) = 
-            abi.decode(cmd, (uint8,bytes32,uint32,address,bytes));
+
+    function resetNonceCond(bytes calldata cmd) private {
+        (, bytes32 salt, uint32 nonce, address oracle, bytes memory args) =
+            abi.decode(cmd, (uint8, bytes32, uint32, address, bytes));
         resetNonceCond(salt, nonce, oracle, args);
     }
 
-    function checkGateOracle (bytes calldata cmd) private {
-        (, address oracle, bytes memory args) = 
-            abi.decode(cmd, (uint8,address,bytes));
+    function checkGateOracle(bytes calldata cmd) private {
+        (, address oracle, bytes memory args) = abi.decode(cmd, (uint8, address, bytes));
         checkGateOracle(oracle, args);
     }
 
@@ -291,11 +299,11 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
      *                permission to.
      * @param nCalls The number of calls the router agent is approved for.
      * @param callpaths The proxy sidecar indexes the router is approved for */
-    function approveRouter (bytes calldata cmd) private {
+    function approveRouter(bytes calldata cmd) private {
         (, address router, uint32 nCalls, uint16[] memory callpaths) =
             abi.decode(cmd, (uint8, address, uint32, uint16[]));
 
-        for (uint i = 0; i < callpaths.length; ++i) {
+        for (uint256 i = 0; i < callpaths.length; ++i) {
             require(callpaths[i] != CrocSlots.COLD_PROXY_IDX, "Invalid Router Approve");
             approveAgent(router, nCalls, callpaths[i]);
         }
@@ -303,8 +311,7 @@ contract ColdPath is MarketSequencer, DepositDesk, ProtocolAccount {
 
     /* @notice Used at upgrade time to verify that the contract is a valid Croc sidecar proxy and used
      *         in the correct slot. */
-    function acceptCrocProxyRole (address, uint16 slot) public pure returns (bool) {
+    function acceptCrocProxyRole(address, uint16 slot) public pure returns (bool) {
         return slot == CrocSlots.COLD_PROXY_IDX;
     }
 }
-
